Moved the project files and folders out of the sub folder and deleted the sub folder
