--- Tic Tac Toe/css/style.css
@@ -1,48 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Alata&family=DM+Sans:ital,opsz,wght@0,9..40,100..1000;1,9..40,100..1000&family=Montserrat:ital,wght@0,100..900;1,100..900&family=Quicksand:wght@300..700&display=swap');
-
-* {
-    font-family: "Quicksand", sans-serif;
-}
-
-.cell {
-    width: 100px;
-    height: 100px;
-    border:none;
-    background-color: none;
-    text-align: center;
-    font-size: 50px;
-}
-
-body {
-    background-image: url("../images/dddepth-345.jpg");
-    background-size: cover;
-    background-repeat: no-repeat;
-}
-
-.border{
-    border-color: #800e149a !important;
-}
-
-.top-left-cell {
-    border-radius: 30px 0 0;
-}
-
-.top-right-cell {
-    border-radius: 0 30px 0 0;
-}
-
-.bottom-left-cell {
-    border-radius: 0 0 0 30px ;
-}
-
-.bottom-right-cell {
-    border-radius: 0 0 30px 0;
-}
-
-.cell-container {
-    background-color: rgba(255, 255, 255, 0.411);
-    backdrop-filter: blur(20px);
-}
-.cell-btn {
-    color: #38040E !important;
-}
